solve data_packing

+++ labs/memory_bound/data_packing/init.cpp
@@ -2,21 +2,36 @@
 #include "solution.h"
 #include <random>
 
+// baseline
+// S create_entry(int first_value, int second_value) {
+//   S entry;
+
+//   entry.i = first_value;
+//   entry.s = static_cast<short>(second_value);
+//   entry.l = static_cast<long long>(first_value * second_value);
+//   entry.d = static_cast<double>(first_value) / maxRandom;
+//   entry.b = first_value < second_value;
+
+//   return entry;
+// }
+
+
 S create_entry(int first_value, int second_value) {
   S entry;
 
-  entry.i = first_value;
+  entry.i = static_cast<short>(first_value);
   entry.s = static_cast<short>(second_value);
-  entry.l = static_cast<long long>(first_value * second_value);
-  entry.d = static_cast<double>(first_value) / maxRandom;
+  entry.l = static_cast<int>(first_value * second_value);
+  entry.d = static_cast<float>(first_value) / maxRandom;
   entry.b = first_value < second_value;
 
   return entry;
 }
 
+
 void init(std::array<S, N> &arr) {
   std::default_random_engine generator;
-  std::uniform_int_distribution<int> distribution(minRandom, maxRandom - 1);
+  std::uniform_int_distribution<int> distribution(minRandom, maxRandom - 1);  // 0, 100
 
   for (int i = 0; i < N; i++) {
     int random_int1 = distribution(generator);
